Rely on python-escpos for low-level commands.
This rips out everything except the layout engine, and makes it run
on python-escpos as the base.
