+++ xmlescpos/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ["constants","escpos","exceptions","printer","supported_devices"]
+from .layout import Layout
